netbox: complete split of generate-inventory-from-netbox.py into modular structure
Split the monolithic NetBox inventory generation script into separate modules:

- config.py: Configuration management
- netbox_client.py: NetBox API client operations
- data_extractor.py: Device data extraction logic
- inventory_manager.py: Inventory file operations
- device_mapping.py: Device role and tag mapping
- dnsmasq_manager.py: dnsmasq configuration management
- utils.py: Shared utility functions

AI-assisted: Claude Code
