+++ files/netbox/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""NetBox integration package for OSISM inventory reconciler."""
