+++ files/netbox/templates/netbox.hosts.j2
